[REFACTOR] 메일 서비스 클래스 추상화
