+++ src/main/java/dorakdorak/infra/mail/service/MailSender.java
@@ -1,0 +1,4 @@
+package dorakdorak.infra.mail.service;
+
+public interface MailSender {
+}
