+++ src/main/java/dorakdorak/infra/mail/service/VerificationMailSender.java
@@ -1,0 +1,7 @@
+package dorakdorak.infra.mail.service;
+
+import jakarta.mail.MessagingException;
+
+public interface VerificationMailSender extends MailSender {
+  boolean sendVerificationMail(String sendEmail) throws MessagingException;
+}
